feat: share internal payment state machine
