+++ modules/fedimint-ln-client/src/lib.rs
@@ -16,12 +16,12 @@ use fedimint_client::module::gen::ClientModuleGen;
 use fedimint_client::module::{ClientModule, IClientModule};
 use fedimint_client::oplog::UpdateStreamOrOutcome;
 use fedimint_client::sm::util::MapStateTransitions;
-use fedimint_client::sm::{Context, DynState, ModuleNotifier, OperationId, State, StateTransition};
+use fedimint_client::sm::{DynState, ModuleNotifier, OperationId, State, StateTransition};
 use fedimint_client::transaction::{ClientOutput, TransactionBuilder};
 use fedimint_client::{sm_enum_variant_translation, Client, DynGlobalClientContext};
 use fedimint_core::api::{DynGlobalApi, DynModuleApi};
 use fedimint_core::config::FederationId;
-use fedimint_core::core::{Decoder, IntoDynInstance, ModuleInstanceId};
+use fedimint_core::core::{IntoDynInstance, ModuleInstanceId};
 use fedimint_core::db::Database;
 use fedimint_core::encoding::{Decodable, Encodable};
 use fedimint_core::module::{
@@ -416,12 +416,6 @@ impl ClientModuleGen for LightningClientGen {
         })
     }
 }
-#[derive(Debug, Clone)]
-pub struct LightningClientContext {
-    pub ln_decoder: Decoder,
-}
-
-impl Context for LightningClientContext {}
 
 #[derive(Debug)]
 pub struct LightningClientModule {
